==> api-ai/ai/ipynb/00.preprocess_faces.ipynb <==
# 데이터 전처리

- 얼굴 데이터 압축 해제 -> 사진 열기 -> 얼굴 검출 -> (224, 224)로 리사이즈


import os
import warnings
from typing import Tuple

import cv2
from preprocess_faces import process_images_mp
from tqdm import tqdm

warnings.filterwarnings(action="ignore")

!ls ./data/Validation

DATA_ROOT_DIR = "./data/Validation"

def detect_faces(img_path: str, dsize: Tuple[int] = (224, 224)):
    if not os.path.exists(img_path):
        return

    img = cv2.imread(img_path)

    # Haar cascade 모델이 흑백 이미지를 사용
    faces = haar_cascade.detectMultiScale(
        cv2.cvtColor(img, cv2.COLOR_BGR2GRAY),
        scaleFactor=1.1,
        minNeighbors=5,
        minSize=(224, 224),
    )

    for x, y, w, h in faces:
        m = max(w, h)
        cv2.rectangle(img, (x, y), (x + m, y + m), (0, 255, 0), 0)
        face_img = cv2.resize(img[y: y + m, x: x + m].copy(), dsize=dsize)

        base, ext = os.path.splitext(img_path)
        out_path = f"face_{base}{ext}"

        cv2.imwrite(out_path, face_img)

    if os.path.exists(img_path):
        os.remove(img_path)

def process_images(path):
    if not os.path.exists(path):
        return

    files = os.listdir(path)

    for fname in tqdm(files, total=len(files)):
        base, ext = os.path.splitext(fname)
        if base[:5] == "face_":
            continue

        detect_faces(os.path.join(path, fname))

for root, dirs, files in os.walk(DATA_ROOT_DIR):
    print(root)

    if len(files) == 0:
        continue

    for fname in files:
        name, ext = os.path.splitext(fname)

        if ext != ".zip":
            continue

        zip_fname = os.path.join(root, fname)
        unzip_dir = os.path.join(root, name)

        if not os.path.exists(unzip_dir):
            unzip_cmd = f"unzip -q {zip_fname} -d {unzip_dir}"
            print(f"Start unzip {fname}")
            os.system(unzip_cmd)

        process_images_mp(unzip_dir)

!df -h
