==> api-ai/ai/ipynb/10.preprocess_answers.ipynb <==
# KoBERT 모델 Fine Tunning을 위한 면접 데이터 전처리


import glob
import json
import os
from pprint import pprint

import pandas as pd
from tqdm import tqdm

USER_HOME = os.path.expanduser("~")
DATA_HOME = os.path.join(
    USER_HOME, "jupyter/129.채용면접 인터뷰 데이터/01-1.정식개방데이터"
)
DATA_TRAINING = os.path.join(DATA_HOME, "Training")
DATA_VALIDATION = os.path.join(DATA_HOME, "Validation")

os.listdir(DATA_HOME)

for d in [DATA_TRAINING, DATA_VALIDATION]:
    zip_list = glob.glob(os.path.join(d, "02.라벨링데이터/*.zip"))

    for src in tqdm(zip_list):
        dest = src.replace(".zip", "")

        if os.path.exists(dest):
            continue

        try:
            os.system(f"unzip -q '{src}' -d '{dest}' > /dev/null 2>&1")
        except Exception as e:
            print(e)

flist = {}
for d in os.listdir(DATA_HOME):
    flist[d] = glob.glob(os.path.join(DATA_HOME, d, "*/*/*.json"))

print(flist.keys())

flist = glob.glob(os.path.join(DATA_HOME, "*/*/*/*.json"))
print(len(flist))
pprint(flist[:3])

### 데이터 어떻게 생겼는지 확인


for fname in tqdm(flist[:3]):
    with open(fname, "r") as f:
        data = json.load(f)["dataSet"]["answer"]
        pprint(data)

## 파일에서 읽고 DataFrame 생성


emotion_answers = {
    "Training": [],
    "Validation": [],
}
intent_answers = {
    "Training": [],
    "Validation": [],
}

for fname in tqdm(flist, total=len(flist)):
    kind = fname.replace(DATA_HOME + "/", "").split("/")[0]

    with open(fname, "r") as f:
        try:
            data = json.load(f)["dataSet"]["answer"]

            # [text, category, expression]
            for e in data["emotion"]:
                if e["text"] == "":
                    continue
                emotion_answers[kind].append(
                    [e["text"], e["category"], e["expression"]]
                )

            for e in data["intent"]:
                if e["text"] == "":
                    continue
                intent_answers[kind].append(
                    [e["text"], e["category"], e["expression"]])
        except Exception as e:
            print(e)

emotion_answers_df = {}
for k, v in emotion_answers.items():
    emotion_answers_df[k] = pd.DataFrame(
        v, columns=["document", "category", "expression"]
    )

emotion_answers_df

intent_answers_df = {}
for k, v in intent_answers.items():
    intent_answers_df[k] = pd.DataFrame(
        v, columns=["document", "category", "expression"]
    )

intent_answers_df

## 카테고리 정수로 변경


emotion_answers_df["Training"].groupby(by=["category", "expression"]).size()

emotion_labels = (
    emotion_answers_df["Training"]
    .groupby(by=["category", "expression"])
    .size()
    .to_frame()
    .reset_index()
    .drop(columns=[0])
)
emotion_labels["label"] = emotion_labels.index

emotion_labels

pd.merge(emotion_answers_df["Training"], emotion_labels, on=["category", "expression"])

intent_answers_df["Training"].groupby(by=["category", "expression"]).size()

intent_labels = (
    intent_answers_df["Training"]
    .groupby(by=["category", "expression"])
    .size()
    .to_frame()
    .reset_index()
    .drop(columns=[0])
)
intent_labels["label"] = intent_labels.index

intent_labels

## TSV로 내보내기


for k, v in emotion_answers_df.items():
    v = pd.merge(v, emotion_labels, on=["category", "expression"])
    v = v.drop(columns=["category", "expression"])
    v.to_csv(
        f"./data/emotion_answers_{k}.tsv", sep="\t", index_label="id", encoding="utf8"
    )

emotion_labels.to_csv(
    "./data/emotion_labels.tsv", sep="\t", index=False, encoding="utf8"
)

for k, v in intent_answers_df.items():
    v = pd.merge(v, intent_labels, on=["category", "expression"])
    v = v.drop(columns=["category", "expression"])
    v.to_csv(
        f"./data/intent_answers_{k}.tsv", sep="\t", index_label="id", encoding="utf8"
    )

intent_labels.to_csv("./data/intent_labels.tsv", sep="\t", index=False, encoding="utf8")

## 파일 확인


read_test_df = pd.read_csv(
    "./data/emotion_answers_Training.tsv", sep="\t", encoding="utf8"
)
read_test_df[:3]
