==> api-ai/ai/ipynb/12.inference.ipynb <==
# 면접 답변 의도 분석


# AttributeError: module 'numpy' has no attribute 'bool'.
# https://stackoverflow.com/questions/74893742/how-to-solve-attributeerror-module-numpy-has-no-attribute-bool
import numpy as np

np.bool = np.bool_

import os

import gluonnlp as nlp
import pandas as pd
import torch
from kobert import BERTDataset
from kobert_tokenizer import KoBERTTokenizer
from tqdm.notebook import tqdm

PRETRAINED = "skt/kobert-base-v1"

tokenizer = KoBERTTokenizer.from_pretrained(PRETRAINED)
vocab = nlp.vocab.BERTVocab.from_sentencepiece(
    tokenizer.vocab_file, padding_token="[PAD]"
)

os.environ["CUDA_DEVICE_ORDER"] = "PCI_BUS_ID"
os.environ["CUDA_VISIBLE_DEVICES"] = "4"

device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
device

dataset_test = pd.read_csv(
    "./data/intent_answers_Validation.tsv", sep="\t", encoding="utf8"
)

dataset_test[:3]

model = torch.load("./models/KoBERT/KoBERT.pt")
model

tok = tokenizer
tok

# Setting parameters
max_len = 128
batch_size = 64

def predict(sentence):
    dataset = [[sentence, "0"]]
    test = BERTDataset(dataset, 0, 1, tok, vocab, max_len, True, False)
    test_dataloader = torch.utils.data.DataLoader(
        test, batch_size=batch_size, num_workers=2
    )

    model.eval()

    answer = 0

    for batch_id, (token_ids, valid_length, segment_ids, label) in enumerate(
        test_dataloader
    ):
        token_ids = token_ids.long().to(device)
        segment_ids = segment_ids.long().to(device)
        valid_length = valid_length
        label = label.long().to(device)
        out = model(token_ids, valid_length, segment_ids)
        for logits in out:
            logits = logits.detach().cpu().numpy()
            answer = np.argmax(logits)
    return answer

intent_labels = pd.read_csv("./data/intent_labels.tsv", sep="\t", encoding="utf8")
intent_labels

acc = 0

for idx, row in tqdm(dataset_test.iterrows(), total=len(dataset_test)):
    sentence = row["document"]
    p = predict(sentence)
    if idx < 3:
        print(f"{sentence = }")
        print(f"predict: {intent_labels.iloc[p][['category','expression']].to_dict()}")
        print(f"actual: {intent_labels.iloc[row['label']][['category','expression']].to_dict()}")
        print()
    acc += 1 if p == row["label"] else 0

print(f"{acc} / {len(dataset_test)} = { 100 * (acc / len(dataset_test))}")
